requirements.txt
streamlit
numpy
pandas
scikit-learn
scipy
matplotlib
seaborn
joblib
torch
fair-esm
xgboost
tqdm
